+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # Unreleased
 
+* Adding AppDeployPipeline construct
+
 # 3.17.3
 
 * adding excluded_instance_families and excluded_instance_sizes to asg